chore(currencies): extract/rename mapCurrencies to mapCurrenciesAsSelectOptions
